// ui/lib/api/schemas/bootstrapConfig.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */
import type { BootstrapHttpConfig } from './bootstrapHttpConfig';
import type { BootstrapPathConfig } from './bootstrapPathConfig';

export interface BootstrapConfig {
  http: BootstrapHttpConfig;
  models: BootstrapPathConfig;
  runtime: BootstrapPathConfig;
}